+++ src/app/ui/catalog/catalog.component.spec.ts
@@ -1,10 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CatalogComponent } from './catalog.component';
-import {ItemService} from './item.service';
-import {OrderService} from '../order/order.service';
-import {CustomerService} from '../member/customer.service';
-import {MerchantService} from '../member/merchant.service';
+import {ItemService} from '../../model/item/item.service';
+import {OrderService} from '../../model/order/order.service';
+import {CustomerService} from '../../model/member/customer.service';
+import {MerchantService} from '../../model/member/merchant.service';
 
 describe('CatalogComponent', () => {
   let component: CatalogComponent;
