+++ src/app/ui/order/order.component.html
@@ -35,7 +35,11 @@
     <td mat-cell *matCellDef="let element">{{element.status}}</td>
     <td mat-footer-cell *matFooterCellDef></td>
   </ng-container>
-
+  <ng-container matColumnDef="action">
+    <th mat-header-cell *matHeaderCellDef></th>
+    <td mat-cell *matCellDef="let element">{{element.status}}</td>
+    <td mat-footer-cell *matFooterCellDef></td>
+  </ng-container>
 
   <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
   <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
